\name{sharx-package}
\alias{sharx-package}
\alias{sharx}
\docType{package}
\title{
Models and Data Sets for the Study of the Species-Area Relationships
}
\description{
Data sets and SAR, SARX, HSAR and HSARX 
models as described in Solymos and Lele (in press).
}
\details{
See \code{\link{sardata}} and \code{\link{hsarx}}
}
\author{
Peter Solymos

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Solymos, P. and Lele, S. R. (in press), Global pattern and local variation
in species-area relationships. 
Global Ecology and Biogeography, xx: xx--xx.
}
\keyword{ package }
