\name{sharx-package}
\alias{sharx-package}
\alias{sharx}
\docType{package}
\title{
Models and Data Sets for the Study of the Species-Area Relationships
}
\description{
Data sets and SAR, SARX, HSAR and HSARX 
models as described in Solymos and Lele (2012).
}
\details{
See \code{\link{sardata}} and \code{\link{hsarx}}
}
\author{
Peter Solymos

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Solymos, P. and Lele, S. R., 2012. 
Global pattern and local variation in species-area relationships. 
\emph{Global Ecology and Biogeography} 21, 109--120.
}
\keyword{ package }
