\name{sharx-package}
\alias{sharx-package}
\alias{sharx}
\docType{package}
\title{
Models and Data Sets for the Study of the Species-Area Relationships
}
\description{
Data sets and SAR, SARX, HSAR and HSARX 
models as described in Solymos and Lele (under review).
}
\details{
See \code{\link{sardata}} and \code{\link{hsarx}}
}
\author{
Peter Solymos

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Solymos and Lele (under review), Global pattern and local variation
in species-area relationship
}
\keyword{ package }
